Correct caps negotiation issue in all elements.
Tracker ID: 197

All of the elements were using gst_pad_proxy_getcaps() as their "getcaps"
function, but this is not the correct behavior.  This was intersecting the
source and sink pad capabilities and returning "EMPTY" as the available caps
during negotiation.

The correct behavior for our elements is to fixate on the static template
caps.

This bug fix allows us to remove the patch we've been using to revert a change
in GStreamer 0.10.13 that forces more strict caps checking.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@38 24075187-2e39-4e88-bbb8-bc8aa768f540