Place the plugin registry cache in /tmp.
This will allow GStreamer to run correctly on systems where the root file
system is mounted read-only.  Most devices use a ramdisk for /tmp, so this
should be safe.

This issue was exposed while trying to run a GStreamer demo off of an SD car
on OMAP 35xx, where the SD card contained the device file system and was
mounted read-only.

Some comments were also added to target_env.sh to explain the purpose of the
environment variables being set.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@40 24075187-2e39-4e88-bbb8-bc8aa768f540