Optimize circular buffer for encoding with support for fixed-sized blocks.
When doing encode, we need to optimize for the case where we know the codec
will alwas consume the same amount of data.  Otherwise, we will spend a lot
of time doing unnecessary memcpys of entire frames.

With this change, gst_ticircbuffer_new takes a new argument named
"fixedBlockSize".  If this is set to true, the circular buffer will set the
readAheadSize to 0, and will only reset the read and write pointer when they
reach the end of the buffer, and they will always be reset the the buffer's
beginning.

From the consumer (codec) point of view, this effectively turns the circular
buffer into a glorified array of fixed-sized buffers.  However, we still have
the ability to accept input buffers in smaller pieces, which might be the
case if we're encoding from a file or network source.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@55 24075187-2e39-4e88-bbb8-bc8aa768f540