Fix build issue with ALSA sink on MontaVista 5.0 toolset.
Point CROSS_COMPILE_PREFIX to the target filesystem that is in the MontaVista
installation directory, not the one you are using for your custom target
filesystem.  The reason is for libraries like libpthread.so.  If you look at
the contents of libpthread.so, you'll find that it isn't really a library, but
a linker script:

---------------------------------------------------------------
$ cat $(TARGET_ROOT_DIR)/usr/lib/libpthread.so

/* GNU ld script
   Use the shared library, but some functions are only in
   the static library, so try that secondarily.  */
OUTPUT_FORMAT(elf32-littlearm)
GROUP ( /lib/libpthread.so.0 /usr/lib/libpthread_nonshared.a )
---------------------------------------------------------------

Notice that it references two other libraries using absolute paths.  If the
MontaVista linker links in libpthread.so from within its install dir, it is
smart enough to know that "/lib/libpthread.so.0" really means a library on the
target filesystem.  However, if it links in libpthread.so from some other
location, it literally interprets "/lib/libpthread.so.0" as a library on the
host filesystem, which will result in a link error.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@145 24075187-2e39-4e88-bbb8-bc8aa768f540