Fix memory leak when flushing the Fifo
This patch fixes a memory leak that can occur when flushing the Fifo.  By
pushing the gst_ti_flush_fifo buffer to the Fifo we are waking up any
processes stuck in a Fifo_get call.  We can then free all the buffers on
the Fifo until we get the gst_ti_flush_fifo buffer.  At this point we know
the Fifo is empty and can be deleted.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@171 24075187-2e39-4e88-bbb8-bc8aa768f540