Correct the number of channels setting when playing elementary streams.
When playing elementary audio streams, the number of channels gets set to 0
because there is no caps negotiation.  This causes a divide by zero exception
when calculating the amount of elapsed play time.

With this change, the number of channels defaults to 2 with a warning when it
could not be determined, and a "numChannels" property was added so it could be
set explicitly by the application in this case.

Tracker ID:  703

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@218 24075187-2e39-4e88-bbb8-bc8aa768f540