Correct circular buffer error check when in fixedBlockSize mode.
Previously it was generating an error whenever we reported that fewer than
"windowSize" bytes were consumed in fixedBlockSize mode.  However, it is valid
to have a zero-length buffer at the end of the stream, so consuming a
zero-length buffer is ok in this case.  The error check has been modified to
make sure the buffer is the same size as the one given out when in
fixedBlockSize mode, which is more correct.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@223 24075187-2e39-4e88-bbb8-bc8aa768f540