Update project information for MPEG2 TS demuxer.
This is a follow-up commit to trunk@42.  In that commit, we removed the
Fluendo MPEG2 TS demuxer open-source package because it had been moved to
gst-plugins-bad, and we have been using it from there ever since.  This commit
updates the README.TXT to reflect this, which we failed to do when we did the
upgrade.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@262 24075187-2e39-4e88-bbb8-bc8aa768f540