Correct free memory read during error check.
This is a follow-up commit to trunk@234.  I modified an error check that used
to compare against "0" to instead compare against GST_BUFFER_SIZE(buf).
Unfortunately, "buf" had already been unref'd by this point in time causing
a free memory read.  This caused the error message to trigger in error
sometimes.

This commit moves the error check before the unref of "buf".

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@263 24075187-2e39-4e88-bbb8-bc8aa768f540