Update README.TXT to reflect the upgrade to DVSDK 2.00 for supported platforms.
DM6446, DM355, and DM6467 have been upgraded to build and run with DVSDK 2.00.
This change primarily affects the gstreamer_demo scripts to accommodate updated
codecs.  The only change to the plugin code itself affects the auto-detection of
the AAC decoder on DM6467 to handle the re-naming of the AAC decoder from
"aacdec" to "aachedec" in DVSDK 2.00.

Release 1.00 will be the final release that supports DVSDK 1.30 for DM6446 and
DM355, and DVSDK 1.40 for DM6467.  Future releases may still work for these
DVSDK versions if the gstreamer_demo directory is taken from the 1.00 release
and the name of the AAC decoder is reverted for DM6467.  However, note that
older DVSDK release are no longer actively supported, and over time the
mainline code base will continue to evolve, possibly making it more difficult
for older DVSDKs to be used.

Tracker ID:  780

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@270 24075187-2e39-4e88-bbb8-bc8aa768f540