Add NM=$(MVTOOL_PREFIX)nm to all configure commands.
Many recent versions of the open-source packages exhibit a common build error:

../libtool: eval: line 950: syntax error near unexpected token `|'

From threads I've read, it seems like this is likely due to libtool expecting
the "nm" utility to have a name like "i686-pc-linux-nm" when that executable
doesn't exist.  As a result, it tries to use a method to dump symbols that
isn't supported on Linux, resulting in the error above.

Overriding the "NM" variables to point to the cross-compiler's nm utility
resolves this issue.

Tracker ID:  773

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@275 24075187-2e39-4e88-bbb8-bc8aa768f540