Remove the decodeDrained and encodeDrained variables, which were unnecessary.
These variables were being used to determine if we should wait for the codec
thead to finish draining.  When we needed to wait, we would call
Rendezvous_meet().  However, upon completion of the drain operation the thread
would always call Rendezvous_force() all the time, so it is safe to call
Rendezvous_meet() all the time without checking the value of decodeDrained.
This renders the variables unnecessary, and they were removed.

Tracker ID:  788

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@283 24075187-2e39-4e88-bbb8-bc8aa768f540