Add DM365 support.
This includes changes to the TI GStreamer elements
as well as scripts for testing the DM365 support.

This patch was submitted by Sajesh Saran.

Currently only audio and video decode support is
enabled.  Video decode uses the video accelerators of
the chip.  The audio decode uses the ARM elements such
as mad.  The example audio decode scripts are in place
to demonstrate how to use elements such as mad on this
platform.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@307 24075187-2e39-4e88-bbb8-bc8aa768f540