Delay calling codec_stop during shut-down.
The call to codec_stop frees the circular buffer, so we must guarantee that
the circular buffer will not be accessed by another thread before completely
shutting down the codec thread.  We do this by blocking in the codec thread's
exit code until either we know the pipeline has started draining (normal case),
or until the main thread is trying to shut down the codec thread (abnormal
termination).

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@365 24075187-2e39-4e88-bbb8-bc8aa768f540