Added gst_ti_calc_buffer_size() function.
Calculate the size of a buffer based on its dimensions and colorspace.  This
function is based on the DMAI function called BufferGfx_calcSize, but it takes
a video standard instead of dimensions.  You don't really need to know the
video standard to calculate the size (some of which only vary in framerates),
and in some cases the video standard is not known.

It would be good if this function could be added to DMAI and the current
"BufferGfx_calcSize" function just calls it using the dimensions from the video
standard.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@388 24075187-2e39-4e88-bbb8-bc8aa768f540