Set parent_class in gst_tidmaivideosink_class_init.
As far as I can tell, parent_class was only used once in the display sink,
but until now it has been left uninitialized.  The only place it was used
was to pass a state change to the base class, and this may not have been
done correctly until now.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@401 24075187-2e39-4e88-bbb8-bc8aa768f540