Only adjust display buffer colorspace for non-DMAI input buffers.
This is a follow-up commit to trunk@376.  On DM365, when getting input from
v4l2src there may be padding on the end of each line, so we need to adjust
the pitch of the input buffers to reflect this.  This is difficult, because
we need to detect this scenario specifically so we don't accidentally adjust
the pitch from elements like videotestsrc and the video decoders.

The best we can do right now is to check the colorspace, and see if the size
of the buffer is larger than it should be.  In this case, we are assuming that
the pitch needs to be adjusted.  Further, we only do this when the input buffer
is not a DMAI transport buffer.  It is ok to assume that if the input was a
DMAI transport buffer, than the creator of the buffer would have set the
buffer's pitch correctly.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@403 24075187-2e39-4e88-bbb8-bc8aa768f540