Move caps processing code into gst_tidmaivideosink_process_caps().
This change is in preperation of supporting downstream buffer allocation by the
TIDmaiVideoSink.  Previously, caps information was read directly by the render
function.  However, when we add the ability to allocate buffers for upstream
elements, we may need to process caps information before the render function
receives its first input buffer.  In this case, it makes sense to abstract the
caps reading code into its own function so it can be shared by the two
use-cases.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@415 24075187-2e39-4e88-bbb8-bc8aa768f540