Add new method for displaying frames multiple times.
Until now, if we have a 60Hz display and we are playing a 30fps clip, we will
copy and display each frame twice.  Unfortunately, on DM365 we don't have the
bandwidth to copy a frame multiple times to display them more than once.
We must put a frame to the display once and let it be refreshed multiple times.

Care must be taken in doing this, as the display driver will return a buffer to
us after refreshing it only once.  If we intend for it to be refreshed multiple
times, we need to make sure we don't write anything into a display buffer while
the display driver could potentially be reading it again for another refresh.
This can be avoided by waiting for a little more than the length of a refresh
cycle after retrieving a display buffer.

This change adds a "repeat_with_refresh" property to the DMAI Video Sink.  If
this field is true, the video sink will block for framerepeat-1 refresh cycles
immediately after retrieving a buffer from the display.

This accomplishes two things:

1)  It throttles the display to the desired framerate.
2)  It avoids ghosting or other glitches that could occur if we wrote into it
    while the display driver is in the middle of reading it again for a
    second refresh cycle.

This feature could potentially be used on DM6467 as well, but I have not looked
to see if it would improve performance or make it worse on DM6467.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@417 24075187-2e39-4e88-bbb8-bc8aa768f540