Fix issue with recent Makefile change.
Recent change to build system (trunk@423) was allowing the plugin libraries to
be built in parallel with the base libraries.  This won't work as the plugins
depend on the base libraries, and the base libraries need to be built first.
This commit re-asserts that the plugins depend on the base libraries.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@424 24075187-2e39-4e88-bbb8-bc8aa768f540