Fix circular buffer issue when input buffers are larger than a window.
It was noticed that AAC playback was hanging (tracker 959).  The circular
buffer was getting stuck becuse it couldn't properly handle the case where
an input buffer was larger than a window.

This check-in modifies the circular buffer to always copy part of the input
buffer if there was space available.  This was previously only done for encode.
Also, it no longer assumes that calls to gst_ticircbuffer_shift_data() will
always succeed.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@459 24075187-2e39-4e88-bbb8-bc8aa768f540