Move call to exit_video in set_sink_caps to the beginning.
As a precaution, we call exit_video inside of set_sink_caps to make sure we
clean-up between clips.  However, the intention was that it was called early-on
or else you risk un-initializing anything that set_sink_caps initializes.  This
change moves the call to set_sink_caps back to the beginning of the function.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@464 24075187-2e39-4e88-bbb8-bc8aa768f540