Initialize videnc1->hCpu and videnc1->device.
Not initializing videnc1->device was causing problems.  This value is only set
during caps negotiation, so if caps negotiation is not performed (i.e. reading
from file), it never gets set and has a garbage value.  In cases where it was
left uninitialized, the garbage in the variable is typically "0", which makes
it look like it was set to DM355 (in DMAI, 0 == DM355).  This easily by-passes
error-checks for valid device values.

A suggestion is being made to the DMAI team that the value "0" should probably
be reserved as an "invalid" device type.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@465 24075187-2e39-4e88-bbb8-bc8aa768f540