Set videnc1->device in the init function.
The code_start function always looks at videnc1->device when initializing the
codec, but videnc1->device was only being set when caps negotiation was taking
place.  If there was no caps negotiation, it was not set properly and the
codec was failing to initialize.

An example where caps negotiation does not take place:

Read from a file and set all codec parameters via TIVidenc1 properties:

$ gst-launch -v filesrc location=input.yuv ! TIVidenc1              \
    codecName=mpeg4enc engineName=encode contiguousInputFrame=FALSE \
    iColorSpace=UYVY resolution=320x240 frameRate=30 !              \
    filesink location=output.m4v

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@467 24075187-2e39-4e88-bbb8-bc8aa768f540