Remove demo scripts from gstreamer_demo.
With all of the additional platforms and use-cases being explored with the
TI GStreamer plugin, the demo scripts (decode_avi.sh, encode_elementary.sh,
etc.) were evolving to a point of having a long list of options and syntaxes
that were becoming fairly complex.  If plugin users need to learn a new syntax,
we would prefer that they learn the syntax of gst-launch itself instead of
that of some demo scripts.

We have removed the demo scripts in favor of a wiki page that documents the
same examples using the native gst-launch syntax.  This wiki page is located
at:

    http://wiki.davincidsp.com/index.php/Example_GStreamer_Pipelines

All that remains in the gstreamer_demo directory is the loadmodules.sh
script for each platform (needed to load the appropriate kernel modules), and
an EXAMPLE_PIPELINES.TXT file that points people to the new wiki page.

Other Details
-------------
Another more subtle change with this commit is that gstreamer_demo/shared is
no longer copied onto the target.  Instead, all the files in
gstreamer_demo/shared are copied into whatever target directory is being
installed.  For example, gstreamer_demo/shared/EXAMPLE_PIPELINES.TXT becomes
gstreamer_demo/dm6446/EXAMPLE_PIPELINES.TXT when installing for DM6446, etc.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@481 24075187-2e39-4e88-bbb8-bc8aa768f540