Fix issue where codecs could be fed garbage data.
This issue has most commonly been associated with the following failure
message:

gsttiauddec1.c:1280:gst_tiauddec1_decode_thread: failed to decode audio buffer

It has been noticed during AAC playback, but it is possible that this could
also occur in other places.

The code in the circular buffer that resets the write pointer from the end of
the buffer to the beginning was incorrectly assuming that each input buffer
contained at least one complete encoded frame.  If input buffers don't contain
complete encoded frames we risk splitting a sample or frame between the end of
the circular buffer in the beginning.  When this happens, the codec gets the
first part of the sample or frame, and then whatever garbage was after it in
the buffer.

Here is a picture of the problem scenario:

 Circular Buffer:
 ------------------------------------------------------------------------------
 | W1                     | W2                      | W3                      |
 ------------------------------------------------------------------------------
  -I1--|-I2-----|           |               |-I1----------| (garbage)
  -S1-------|-S2-           |                   |-S1------- (garbage)
                W           R                             W
                ^                                         |
                |-----------------------------------------|

 Wn = Circular Buffer Window
 In = Input Buffer
 Sn = Audio Sample
 R  = Read Pointer
 W  = Write Pointer

As you can see, when S1 is processed, only the first part of the sample will
be valid -- the rest will be garbage.  The read pointer is not reset to W1
until it reaches W3, but S1 starts before W3.

This change makes the code that resets the write pointer more conservative,
and won't reset the write pointer unless it can guarantee that any read
that spills from W2 into W3 will get a complete frame.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@527 24075187-2e39-4e88-bbb8-bc8aa768f540