Fix build issue with check library.
When building the check library on a RHEL4 machine using CodeSourcery 2009q1,
the following error was observed when building the check library:

../lib/.libs/libcompat.a: could not read symbols: Archive has no index; run ranlib to add one
collect2: ld returned 1 exit status

This was resulting from ranlib not getting executed properly on libcompat.a,
as the build was not able to find this tool.  Specifying

RANLIB=$(MVTOOL_PREFIX)ranlib

on the configure line specifies the full path to the ranlib utility provided
by the cross-compiler, which allows the build scripts to find it.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@575 24075187-2e39-4e88-bbb8-bc8aa768f540