Change format of patches for dependent components to match review process.
This check-in changes the format of the patches used for dependent components
to use the format of "git format-patch".  These patches are now consistent
with the patch specification used by the project review process.

Having the patches in this format also makes them easier to maintain, as
they now work well with GIT.


git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@608 24075187-2e39-4e88-bbb8-bc8aa768f540