Add support for new DSPLink directory structure.
Historically DSPLink has contained a "packages" sub-directory that is the XDC
root of the component.  In the latest versions this "packages" directory has
been removed.  The build system now adds a variable LINK_XDC_ROOT to indicate
the XDC root of DSPLink that can be set on a per-target basis, and no longer
assumes it is the same as LINK_INSTALL_DIR.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@611 24075187-2e39-4e88-bbb8-bc8aa768f540