Always return NULL for the base sink's last-buffer property.
This property requires the video sink to keep a reference to the last buffer
displayed, which means it always holds on to two input buffers.  In some
embedded systems this is not practical, and in fact has been known to cause
driver deadlock if the input buffers are capture driver buffers (via v4l2src
always-copy=FALSE).  There are also issues freeing resources at EOS if the sink
is holding onto a buffer whose memory is being managed by upstream elements.

I could see no way to disable this functionality at run-time.  Perhaps the
long-solution for this would be to disable this feature via property.  For now,
this patch disables the feature at build time.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@632 24075187-2e39-4e88-bbb8-bc8aa768f540