Don't error out if DMAI_INSTALL_DIR is set by Rules.make.
Previously the Makefile.external file would error out if DMAI_INSTALL_DIR
wasn't set explicitly -- even if it was defined by the DVSDK's Rules.make.
There was a comment indicating that if your Rules.make set DMAI_INSTALL_DIR to
disable setting its default value of "undefined", which implictly subverts
the error check.

This commit moves the DMAI_INSTALL_DIR error check until after Rules.make
is included.  This way, the error check comprehends if it was set by Rules.make
and no modification to the Makefile is needed in this case.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@639 24075187-2e39-4e88-bbb8-bc8aa768f540