Correct size metadata imgdec1->hInBuf prior to calling the codec.
The imaging decoders may fail if the size of the input buffer is not the same
as the size the codec says it should be.  This change corrects the size
of imgdec1->hInBuf to be the result of Idec1_getInBufSize, which is the size
specified by the codec.

Because the circular buffer's window size is also Idec1_getInBufSize, we are
guaranteed to have at least that much readable data available at the pointer
it returns to us.  The number of bytes in the buffer that the codec should
process is specified by the Buffer_setNumBytesUsed method for the input
buffer, which was communicated by the circular buffer.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@641 24075187-2e39-4e88-bbb8-bc8aa768f540