Add GST_TI_PLATFORM variable to replace previous usage of PLATFORM
Historically we have used the DVSDK's $(PLATFORM) variable to uniquely identify
the EVM we were building for.  With the DVSDK 3.10 Beta release we cannot rely
on this, as this version of the DVSDK sets PLATFORM=dm6467 when building for
DM6467T, and doesn't officially support DM6467.  Hence, with this release we
don't have a way to determine if we are building for DM6467 with DVSDK 2.00 or
DM6467T with DVSDK 3.10.

An incident report has been filed against this compatibility break, and
hopefully it will be resolved soon.  Until then...

This change allows us to distinguish between two different EVMs by giving them
different values for GST_TI_PLATFORM, even if their PLATFORM settings are the
same in the DVSDK's Rules.make.

This change does not add DM6467T support -- it is merely making a prerequisite
build system change.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@649 24075187-2e39-4e88-bbb8-bc8aa768f540