Corrected TI Copyright notices (replaced $year with proper years).
Most of the Copyright notices in the source code say "Copyright $year".  This
was an internal convention we used, and typically replaced $year with the
proper year when making a source code release.

However, since the source code repository is external, we should have real
years in our source files always.  This change puts proper year or year ranges
in the source code.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@678 24075187-2e39-4e88-bbb8-bc8aa768f540