DM365:  Leave the IPIPE open during UYVY capture.
This enables on-the-fly resizing during UYVY capture.

Please note that even the whole patch set for gst-plugins-good was
re-generated, the only patch that actually changed was:

0007-v4l2src-support-NV12-capture-on-DM365-using-the-IPIP.patch

This change removes the code that closes the IPIPE when capturing UYVY.

Initially, we open the IPIPE in chaning mode with capture in case we need to
capture NV12.  The IPIPE is needed in this case to convert UYVY to NV12
on-the-fly, which is the format required by the codecs.

However, we previously closed the IPIPE if we discovered we were capturing
UYVY, as it was not needed for NV12 conversion.  It is actually better to
leave it open.

Using the IPIPE during capture has an additional benefit of allowing on-the-fly
resizing -- so it is better to keep it open for both NV12 and UYVY capture so
this feature is always enabled.  When left enabled, any resolution supported by
the IPIPE resizer can be specified as a capture resolution.

For example, capture and resize to 1024x240:

   gst-launch -v v4l2src always-copy=FALSE ! 'video/x-raw-yuv,         \
       format=(fourcc)UYVY, framerate=(fraction)60/1, width=(int)1024, \
       height=(int)240' ! TIDmaiVideoSink contiguousInputFrame=TRUE    \
       videoStd=720P_60 videoOutput=component

Note that on-the-fly resizing is only supported when the DM365 EVM is booted
with the dm365_imp.oper_mode=0 boot argument.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@679 24075187-2e39-4e88-bbb8-bc8aa768f540