Remove v4l2 patch that resets bytesperline to 0.
A patch for gst-plugins-good/sys/v4l2 was introduced that reset bytesperline to
0 before each call to S_FMT and TRY_FMT.  Most drivers will calculate this
value for the application when it is 0.

Unfortunately, with DM6467, this value needs to be set to the value returned
from G_FMT when initializing the capture device, or it will fail to initialize.

This intended use of this patch was to help enable v4l2sink, but we can't do
this at the cost of breaking v4l2src.  For now, we're reverting this patch
until we can figure out a better way to solve this problem.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@703 24075187-2e39-4e88-bbb8-bc8aa768f540