Correct the way source caps information is generated in TIViddec2.
Before we were hard-coding UYVY for source caps, with one exception for DM365
where we were hard-coding NV12 instead.  With this change, the fourcc is
determined based on the colorspace in the buffer instead of making
target-specific assumptions.

This also fixes an issue on DM6467 where the output caps were reporting UYVY
when they were really NV12.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@704 24075187-2e39-4e88-bbb8-bc8aa768f540