Push all display initialization code into gst_tidmaivideosink_init_display.
As part of the display initialization, the render function performs a few
operations before and after calling gst_tidmaivideosink_init_display.
However, these operations should actually be inside of the
gst_tidmaivideosink_init_display function in case this function ever needs to
be called from other places.

For example, right now we initialize the display only when we get the first
input buffer.  In the future, we may need to initialize the display early, if
we add support for peer-allocated buffers and want to return driver buffers
up-stream.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@707 24075187-2e39-4e88-bbb8-bc8aa768f540