Add a property to hide the OSD on DM365/DM6446/DM355 during video playback.
If hideOSD=TRUE is passed to TIDmaiVideoSink, the OSD is opened, initialized,
and made transparent so it doesn't interfere with video playback.  This is
especially useful when fbcon is enabled in the kernel -- TIDmaiVideo sink will
make sure the fbcon is hidden when displaying video.

If this property is used, it should eliminate the need to run commands like:

cat /dev/zero > /dev/fb2 2> /dev/null

To clear the fb attribute window after booting.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@730 24075187-2e39-4e88-bbb8-bc8aa768f540