Add parameter to gst_ti_calc_buffer_size that specifies pitch (bytes per line).
Previously, gst_ti_calc_buffer_size assumed that the pitch could always be
inferred from the colorspace.  This is not the case on platforms like DM365,
where display buffers may require additional padding.  For example, a 720x480
frame needs a pitch of 736 to satisfy the driver's requirement that the pitch
is 32-byte aligned.  There is no way to comprehend this in a generic way.

This change introduces a new "bytesPerLine" parameter that allows the pitch to
be specified explicitly by the caller.  If "0" is specified for bytesPerLine,
the pitch is inferred from the colorSpace as before.

This change does not affect existing behavior.  All calls to
gst_ti_calc_buffer_size were updated to pass "0" for bytesPerLine to keep
previous behavior.  Subsuequent changes may use this new feature to handle
platforms like DM365 better.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@753 24075187-2e39-4e88-bbb8-bc8aa768f540