Fix race condition in GstTIDmaiBufTab.
This change fixes the following piece of code, which may not work properly due
to a race condition:

In gst_tidmaibuftab_get_buf:
     hFreeBuf = BufTab_getFreeBuf(self->hBufTab);

     while (!hFreeBuf) {
         Rendezvous_meet(self->hBufAvailRv);
         hFreeBuf = BufTab_getFreeBuf(self->hBufTab);
         Rendezvous_reset(self->hBufAvailRv);
     }

This code is supposed to try and get a free buffer, and if it fails, it is
supposed to block until a free buffer is available.  This was not always
working due to a race condition.

To explain the race condition, note that two things occur when a buffer is
made available for re-use:

1)  A call to Buffer_freeUseMask is made that clears a bit.  Once this bit
is cleared, BufTab_getFreeBuf will return it.

2)  A call to Rendezvous_force is made, which will unblock the above call to
Rendezvous_meet, which tells the above code to try again to get a buffer, and
there is really no reason why it should fail this time -- one was just freed
up.

However, sometimes, this race condition occurs:

Thread 1:
     Call Buffer_freeUseMask() to mark a buffer as free for use.

Thread 2:
     hFreeBuf = BufTab_getFreeBuf(self->hBufTab);
     <hFreBuf is non-null, returned, and used>

     hFreeBuf = BufTab_getFreeBuf(self->hBufTab);
     <this time we get NULL because we just used the only buffer>

     while (!hFreeBuf) {
         Rendezvous_meet(self->hBufAvailRv);
     <wait for a buffer>

Thread 1:
     Call Rendezvous_force() for the *first* buffer that has already been
     consumed by Thread 2.

Thread 2:
         hFreeBuf = BufTab_getFreeBuf(self->hBufTab);
         Rendezvous_reset(self->hBufAvailRv);
     }
     <hFreeBuf is still NULL, because we got un-blocked by a stale call to
      Rendezvous_force()>

This change puts pthread mutexes around the calls to BufTab_getFreeBuf, and
grabs the mutex between the calls to Buffer_freeUseMask and Rendezvous_force().
This should solve the race condition and keep the bookeeping in tact.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@782 24075187-2e39-4e88-bbb8-bc8aa768f540